ADD line under mouse and circle on graph CHANGE sensordata type REMOVE unused field

+++ gui/MotorXPGUI/MotorXPGUIMVVM/Model/SensorData.cs
@@ -5,6 +5,6 @@ namespace MotorXPGUIMVVM.Model
     public class SensorData
     {
         public ulong Timestamp { get; set; }
-        public SortedList<ushort, ulong> DataTable { get; set; }
+        public SortedList<ushort, double> DataTable { get; set; }
     }
 }
